refactor(shopify): simplify naming structure
